Break resource analysis specs into its own module
